ZoneManager created
Working class ZoneManager

TODO : assistant
