+++ simpleDNS/class/CrontrolWritter.cs
