+++ simpleDNS/forms/frmAboutBox.resx
