+++ simpleDNS/forms/frmAssistant.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Text;
+using System.Windows.Forms;
+
+namespace simpleDNS
+{
+    public partial class frmAssistant : Form
+    {
+        public frmAssistant()
+        {
+            InitializeComponent();
+        }
+    }
+}
