+++ simpleDNS/forms/frmMain.resx
