+++ simpleDNS/simpleDNS.csproj
@@ -12,10 +12,10 @@
     <PackageProjectUrl>https://github.com/AnthoDingo/simpleDNS</PackageProjectUrl>
     <RepositoryUrl>https://github.com/AnthoDingo/simpleDNS</RepositoryUrl>
     <RepositoryType>GIT</RepositoryType>
-    <AssemblyVersion>1.0.0.21290</AssemblyVersion>
-    <FileVersion>1.0.0.21290</FileVersion>
+    <AssemblyVersion>1.0.0.21291</AssemblyVersion>
+    <FileVersion>1.0.0.21291</FileVersion>
     <Description>Simple DNS à destination de l'éducation.</Description>
-    <SignAssembly>true</SignAssembly>
+    <SignAssembly>false</SignAssembly>
     <AssemblyOriginatorKeyFile>AnthoDingo.pfx</AssemblyOriginatorKeyFile>
   </PropertyGroup>
